+++ Changelog.md
@@ -5,6 +5,10 @@
 6/1/2021 - Hustin
 
 - Moved the container and content div's back into MainLayout, it was causing styling issues being in Routes. Not sure why.
+- Delete reply server/client side
+- Adjusted TextArea inputs for reply to have minRow of 4 and autosize
+- Fixed saved post not showing up for users who didn't post it
+- Edit/Delete reply buttons only show for that users reply
 
 5/31/2021 - Hustin